Tests/LibWeb/Text/expected/textarea-value.txt

1. "PASS"
2. "PASS"
3. "PASS"
4. 4
5. "PASS"
6. "[0x0a]text[0x0a]with[0x0a]carriage[0x0a][0x0a]return[0x0a][0x0a]"
